ImageCaptioning/requirements.txt
Django>=2.2
pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.19.5
tensorflow>=2.8.0 